django_apps/requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pillow==2.6.1
gunicorn==19.1.1
image==1.3.4
wsgiref==0.1.2
